src/main/java/pl/coderslab/charity/category/CategoryService.java
package pl.coderslab.charity.category;

import org.springframework.stereotype.Service;

import javax.transaction.Transactional;
import java.util.List;

@Service
@Transactional
public class CategoryService {
    private final CategoryRepository categoryRepository;

    public CategoryService(CategoryRepository categoryRepository) {
        this.categoryRepository = categoryRepository;
    }

    public List<Category> getAllCategory(){
        return categoryRepository.getAllCategory();
    }

    public void saveNewCategory(Category category) {
        categoryRepository.save(category);
    }
}
